Rename util to rl_util
